src/test/java/seedu/address/model/event/DescriptionTest.java
package seedu.address.model.event;

import static junit.framework.TestCase.assertTrue;
import static org.junit.Assert.assertFalse;

import org.junit.Test;

import seedu.address.testutil.Assert;

public class DescriptionTest {

    @Test
    public void constructor_null_throwsNullPointerException() {
        Assert.assertThrows(NullPointerException.class, () -> new Description(null));
    }

    @Test
    public void constructor_invalidDescription_throwsIllegalArgumentException() {
        String invalidDescription = "";
        Assert.assertThrows(IllegalArgumentException.class, () -> new Description(invalidDescription));
    }

    @Test
    public void isValidDescription() {
        // null description
        Assert.assertThrows(NullPointerException.class, () -> Description.isValidDescription(null));

        // invalid description
        assertFalse(Description.isValidDescription(""));
        assertFalse(Description.isValidDescription(" "));

        // valid description
        assertTrue(Description.isValidDescription("25th Anniversary at Marina Bay Sands"));
        assertTrue(Description.isValidDescription("TODO: Prepare coursework"));
        assertTrue(Description.isValidDescription("-"));
    }
}
